config: exynos850: Set BOARD_USE_CUSTOM_COMPONENT_SUPPORT to true
This flag is needed for correspondence with the OMX setup on all
non-development Samsung devices.

The Linaro Android 10 erd850 device tree sets this to false. But in any
case, I completely doubt anyone will try to build those trees here. All
of Linaro's Android 13 device trees with this flag have it set to true.

Change-Id: I7b0ec3b4851272edc923f43e622b9fccde4003c9